pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "rag-vehicle-api"
version = "1.0.0"
description = "API RAG para selección inteligente de vehículo - Logística agrícola colombiana"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "anthropic>=0.28.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "langchain-text-splitters>=0.2.0",
    "pypdf>=4.2.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.35.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
